requirements.txt
pandas==3.0.0
Pillow==12.1.1
reportlab==4.4.10
pypdf==6.7.0
openai==2.20.0
